+++ correlate/backend/badger.go
@@ -1,8 +1,8 @@
 package backend
 
 import (
-	"encoding/json"
 	"github.com/dgraph-io/badger/v4"
+	"github.com/go-json-experiment/json"
 	"github.com/pkg/errors"
 	"net/url"
 	"time"
